Add component isolation (#[Isolate] attribute) (#7653)
