+++ js/features/index.js
@@ -8,6 +8,7 @@ import './supportLazyLoading';
 import './supportFileUploads';
 import './supportQueryString';
 import './supportLaravelEcho';
+import './supportIsolating';
 import './supportRedirects';
 import './supportNavigate';
 import './supportMorphDom';
