+++ src/Attributes/Isolate.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Livewire\Attributes;
+
+use Livewire\Features\SupportIsolating\BaseIsolate;
+
+#[\Attribute]
+class Isolate extends BaseIsolate
+{
+    //
+}
