+++ src/Features/SupportLazyLoading/BrowserTest.php
@@ -133,136 +133,6 @@ class BrowserTest extends BrowserTestCase
     }
 
     /** @test */
-    public function lazy_requests_are_isolated_by_default()
-    {
-        Livewire::visit([new class extends Component {
-            public function render() { return <<<HTML
-            <div>
-                <livewire:child num="1" />
-                <livewire:child num="2" />
-                <livewire:child num="3" />
-            </div>
-            HTML; }
-        }, 'child' => new #[\Livewire\Attributes\Lazy] class extends Component {
-            public $num;
-            public $time;
-            public function mount() {
-                $this->time = LARAVEL_START;
-            }
-            public function render() { return <<<'HTML'
-            <div id="child">
-                Child {{ $num }}
-
-                <span dusk="time-{{ $num }}">{{ $time }}</span>
-            </div>
-            HTML; }
-        }])
-        ->waitForText('Child 1')
-        ->waitForText('Child 2')
-        ->waitForText('Child 3')
-        ->tap(function ($b) {
-            $time1 = (float) $b->text('@time-1');
-            $time2 = (float) $b->text('@time-2');
-            $time3 = (float) $b->text('@time-3');
-
-            $this->assertNotEquals($time1, $time2);
-            $this->assertNotEquals($time2, $time3);
-        })
-        ;
-    }
-
-    /** @test */
-    public function lazy_requests_are_isolated_by_default_but_bundled_on_next_request_when_polling()
-    {
-        Livewire::visit([new class extends Component {
-            public function render() { return <<<HTML
-            <div>
-                <livewire:child num="1" />
-                <livewire:child num="2" />
-                <livewire:child num="3" />
-            </div>
-            HTML; }
-        }, 'child' => new #[\Livewire\Attributes\Lazy] class extends Component {
-            public $num;
-            public $time;
-            public function boot()
-            {
-                $this->time = LARAVEL_START;
-            }
-
-            public function render() { return <<<'HTML'
-            <div wire:poll.500ms id="child">
-                Child {{ $num }}
-
-                <span dusk="time-{{ $num }}">{{ $time }}</span>
-            </div>
-            HTML; }
-        }])
-        ->waitForText('Child 1')
-        ->waitForText('Child 2')
-        ->waitForText('Child 3')
-        ->tap(function ($b) {
-            $time1 = (float) $b->text('@time-1');
-            $time2 = (float) $b->text('@time-2');
-            $time3 = (float) $b->text('@time-3');
-
-            $this->assertNotEquals($time1, $time2);
-            $this->assertNotEquals($time2, $time3);
-        })
-        // Wait for a poll to have happened
-        ->pause(500)
-        ->tap(function ($b) {
-            $time1 = (float) $b->text('@time-1');
-            $time2 = (float) $b->text('@time-2');
-            $time3 = (float) $b->text('@time-3');
-
-            // Times should now all be equal
-            $this->assertEquals($time1, $time2);
-            $this->assertEquals($time2, $time3);
-        })
-        ;
-    }
-
-    /** @test */
-    public function lazy_requests_can_be_bundled_with_attribute_parameter()
-    {
-        Livewire::visit([new class extends Component {
-            public function render() { return <<<HTML
-            <div>
-                <livewire:child num="1" />
-                <livewire:child num="2" />
-                <livewire:child num="3" />
-            </div>
-            HTML; }
-        }, 'child' => new #[\Livewire\Attributes\Lazy(isolate: false)] class extends Component {
-            public $num;
-            public $time;
-            public function mount() {
-                $this->time = LARAVEL_START;
-            }
-            public function render() { return <<<'HTML'
-            <div id="child">
-                Child {{ $num }}
-
-                <span dusk="time-{{ $num }}">{{ $time }}</span>
-            </div>
-            HTML; }
-        }])
-        ->waitForText('Child 1')
-        ->waitForText('Child 2')
-        ->waitForText('Child 3')
-        ->tap(function ($b) {
-            $time1 = (float) $b->text('@time-1');
-            $time2 = (float) $b->text('@time-2');
-            $time3 = (float) $b->text('@time-3');
-
-            $this->assertEquals($time1, $time2);
-            $this->assertEquals($time2, $time3);
-        })
-        ;
-    }
-
-    /** @test */
     public function can_lazy_load_component_using_route()
     {
         $this->tweakApplication(function() {
